refactored. moved AuthPage.jsx component into /pages directory
